some commented code extracted to readme file.
