+++ F12020Structures/F12020/PacketCarStatusData.cs
@@ -1,0 +1,14 @@
+﻿using System.Runtime.InteropServices;
+
+namespace SimHubToF12020UDP
+{
+
+    [StructLayout(LayoutKind.Sequential, Pack = 1)]
+    public struct PacketCarStatusData
+    {
+        public PacketHeader m_header;           // Header
+
+        [MarshalAs(UnmanagedType.ByValArray, SizeConst = 22)]
+        public CarStatusData[] m_carStatusData;
+    };
+}
